% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{apply_simulations}
\alias{apply_simulations}
\title{Apply Simulation Function}
\usage{
apply_simulations(df, funct)
}
\arguments{
\item{df}{A data frame to be nested and processed.}

\item{funct}{A function to be applied to each nested data frame.}
}
\value{
A data frame with the results of the applied function.
}
\description{
This function applies a given function to a nested data frame with `nest_cols = "simulation"`.
}
