% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{build_correlation_matrix}
\alias{build_correlation_matrix}
\title{Create Correlation Matrix}
\usage{
build_correlation_matrix(
  correlation_data,
  arg = rlang::caller_arg(correlation_data),
  call = rlang::caller_env()
)
}
\arguments{
\item{correlation_data}{A data frame with three columns: effect1, effect2, and correlation.}

\item{arg}{The argument name to use in error messages.}

\item{call}{The environment to use for error messages.}
}
\value{
A correlation matrix.
}
\description{
This function creates a correlation matrix from a data frame with three columns: effect1, effect2, and correlation.
}
