% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calculate_expectation}
\alias{calculate_expectation}
\title{Calculate Expectation}
\usage{
calculate_expectation(
  data,
  model_matrix,
  fixed_effects_coefficients,
  random_effects_sd = NULL,
  random_effects_corr_matrix = list()
)
}
\arguments{
\item{model_matrix}{A matrix representing the model.}

\item{fixed_effects_coefficients}{A vector of fixed effects coefficients.}

\item{random_effects_sd}{A named list of standard deviations for each random effect (default is NULL).}

\item{random_effects_corr_matrix}{A named list of correlation matrices for each random effect (default is an empty list).}

\item{data.}{A data frame with the original data used to generate the model matrix.}
}
\value{
A vector representing the calculated expectation.
}
\description{
This function calculates the expectation of a model given the fixed effects coefficients and random effects.
}
