% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_function}
\alias{check_function}
\title{Check if an object is a function}
\usage{
check_function(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{An object.}

\item{arg}{The argument name.}

\item{call}{The calling environment.}
}
\value{
Throws an error if the object is not a function.
}
\description{
Check if an object is a function
}
