% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_integer_vector}
\alias{check_integer_vector}
\title{Check if a vector contains only integers}
\usage{
check_integer_vector(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{A numeric vector.}

\item{arg}{The argument name.}

\item{call}{The calling environment.}
}
\value{
Throws an error if the vector contains non-integer values.
}
\description{
Check if a vector contains only integers
}
