% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_logical_vector}
\alias{check_logical_vector}
\title{Check if a vector is of type logical}
\usage{
check_logical_vector(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{A vector.}

\item{arg}{The argument name.}

\item{call}{The calling environment.}
}
\value{
Throws an error if the vector is not of type logical.
}
\description{
Check if a vector is of type logical
}
