% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_names_match}
\alias{check_names_match}
\title{Check if Names in a Vector Match Names in a Data Frame}
\usage{
check_names_match(
  vec,
  df,
  check_order = FALSE,
  vec_arg = rlang::caller_arg(vec),
  df_arg = rlang::caller_arg(df),
  call = rlang::caller_env()
)
}
\arguments{
\item{vec}{A named vector whose names are to be checked.}

\item{df}{A data frame whose column names are to be checked against the vector names.}

\item{check_order}{A logical value indicating whether to check if the order of names in the vector matches the order in the data frame. Default is FALSE.}

\item{vec_arg}{The argument name for the vector, used for error messaging. Default is the caller's argument name for `vec`.}

\item{df_arg}{The argument name for the data frame, used for error messaging. Default is the caller's argument name for `df`.}

\item{call}{The environment to use for error messaging. Default is the caller's environment.}
}
\value{
This function does not return a value. It is used for its side effect of throwing an error if the names do not match.
}
\description{
This function checks if all names in a given vector are present in a data frame.
Optionally, it can also check if the order of names in the vector matches the order in the data frame.
}
