% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_positive_values}
\alias{check_positive_values}
\title{Check if a vector contains only positive values}
\usage{
check_positive_values(
  x,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{A numeric vector.}

\item{arg}{The argument name.}

\item{call}{The calling environment.}
}
\value{
Throws an error if the vector contains non-positive values.
}
\description{
Check if a vector contains only positive values
}
