% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_probability_vector}
\alias{check_probability_vector}
\title{Check if a vector is a valid probability vector}
\usage{
check_probability_vector(
  x,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{A numeric vector.}

\item{arg}{The argument name.}

\item{call}{The calling environment.}
}
\value{
Throws an error if the vector is not a valid probability vector.
}
\description{
Check if a vector is a valid probability vector
}
