% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_range_vector}
\alias{check_range_vector}
\title{Check if a vector contains values within a specific range}
\usage{
check_range_vector(
  x,
  min,
  max,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{A numeric vector.}

\item{min}{The minimum value.}

\item{max}{The maximum value.}

\item{arg}{The argument name.}

\item{call}{The calling environment.}
}
\value{
Throws an error if the vector contains values outside the specified range.
}
\description{
Check if a vector contains values within a specific range
}
