% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_unique_names}
\alias{check_unique_names}
\title{Check if a vector has unique names}
\usage{
check_unique_names(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{A named vector.}

\item{arg}{The argument name.}

\item{call}{The calling environment.}
}
\value{
Throws an error if the vector does not have unique names.
}
\description{
Check if a vector has unique names
}
