% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_unique_values}
\alias{check_unique_values}
\title{Check if a vector contains unique values}
\usage{
check_unique_values(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{A vector.}

\item{arg}{The argument name.}

\item{call}{The calling environment.}
}
\value{
Throws an error if the vector contains duplicate values.
}
\description{
Check if a vector contains unique values
}
