% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cov_matrix_from_corr}
\alias{cov_matrix_from_corr}
\title{Compute Covariance Matrix from Correlation Matrix}
\usage{
cov_matrix_from_corr(sds, corr_matrix = NULL)
}
\arguments{
\item{sds}{Numeric vector of standard deviations.}

\item{corr_matrix}{Numeric matrix of correlations.}
}
\value{
Numeric covariance matrix.
}
\description{
Computes the covariance matrix from standard deviations and a correlation matrix.
}
