% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_cartesian_product}
\alias{create_cartesian_product}
\title{Create Cartesian Product}
\usage{
create_cartesian_product(..., .times = 1)
}
\arguments{
\item{...}{Vectors or lists to create the Cartesian product from.}

\item{.times}{Integer. The number of times to replicate each row. Default is 1.}
}
\value{
A data.table containing the Cartesian product of the input vectors or lists, with each row replicated .times times.
}
\description{
This function creates a Cartesian product of the input vectors or lists and replicates each row a specified number of times.
}
