% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{df_apply_nested}
\alias{df_apply_nested}
\title{Nest Data and Apply Function to Each Nested Data Frame}
\usage{
df_apply_nested(
  data,
  nest_cols,
  func,
  result_col = "result",
  data_col = "data",
  startup_function = NULL,
  parallel = FALSE,
  cores = 1
)
}
\arguments{
\item{data}{A data frame to nest.}

\item{nest_cols}{Columns to nest by.}

\item{func}{A function to apply to each nested data frame.}

\item{result_col}{The name of the new column to store the results.}

\item{startup_function}{An optional function to run before processing starts.}

\item{parallel}{A logical value indicating whether to run the processing in parallel.}

\item{cores}{An integer specifying the number of cores to use for parallel processing.}

\item{complete_function}{An optional function to run after each nested data frame is processed.}
}
\value{
A data frame with the results stored in a new column.
}
\description{
This function nests the data based on specified columns, applies a specified function to each nested data frame, and stores the results in a new column.
}
