% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{factor2matrix}
\alias{factor2matrix}
\title{Convert Factor to Matrix}
\usage{
factor2matrix(data, fct)
}
\arguments{
\item{data}{A data frame containing the factor column.}

\item{fct}{The name of the factor column to be converted.}
}
\value{
A matrix with one-hot encoding of the factor column.
}
\description{
This function converts a factor column in a data frame to a matrix using one-hot encoding.
}
