% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{generate_response_binomial}
\alias{generate_response_binomial}
\title{Generate Binomial Response}
\usage{
generate_response_binomial(eta, weights = 1, link = "logit")
}
\arguments{
\item{eta}{Numeric vector of linear predictors.}

\item{weights}{Numeric value of weights.}

\item{link}{Character string specifying the link function.}
}
\value{
Numeric vector of response values.
}
\description{
Generates the response variable for a binomial model with a specified link function.
}
