% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{generate_response_gamma}
\alias{generate_response_gamma}
\title{Generate Gamma Response}
\usage{
generate_response_gamma(eta, dispersion = 1, link = "log")
}
\arguments{
\item{eta}{Numeric vector of linear predictors.}

\item{dispersion}{Numeric value of dispersion parameter.}

\item{link}{Character string specifying the link function.}
}
\value{
Numeric vector of response values.
}
\description{
Generates the response variable for a Gamma model with a specified link function.
}
