% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{generate_response_gaussian}
\alias{generate_response_gaussian}
\title{Generate Gaussian Response}
\usage{
generate_response_gaussian(mu, sigma = 1)
}
\arguments{
\item{mu}{Numeric vector of means.}

\item{sigma}{Numeric value of standard deviation.}
}
\value{
Numeric vector of response values.
}
\description{
Generates the response variable for a linear model with Gaussian errors.
}
