% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{lapply_rowwise}
\alias{lapply_rowwise}
\title{Apply a function to each row of a data frame}
\usage{
lapply_rowwise(
  data,
  func,
  startup_function = NULL,
  complete_function = NULL,
  seed = 123,
  parallel = FALSE,
  cores = 1
)
}
\arguments{
\item{data}{A data frame to process.}

\item{func}{A function to apply to each row of the data frame.}

\item{startup_function}{An optional function to run before processing starts.}

\item{complete_function}{An optional function to run after each row is processed.}

\item{seed}{An integer seed for reproducibility.}

\item{parallel}{A logical value indicating whether to run the processing in parallel.}

\item{cores}{An integer specifying the number of cores to use for parallel processing.}
}
\value{
A list of results from applying the function to each row.
}
\description{
This function applies a specified function to each row of a data frame, with options for parallel processing and additional startup and completion functions.
}
