% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{make_random_effect_matrix}
\alias{make_random_effect_matrix}
\title{Calculate Random Effect Matrix}
\usage{
make_random_effect_matrix(group_factor, random_slopes, corr_matrix)
}
\arguments{
\item{group_factor}{A factor that defines the groups.}

\item{random_slopes}{A numeric vector of standard deviations for each random slope.}

\item{corr_matrix}{A correlation matrix.}
}
\value{
A matrix of random effects for each group.
}
\description{
This function calculates a random effect matrix for each group defined by a factor in the data frame.
}
