% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{model_response}
\alias{model_response}
\title{Simulate Model}
\usage{
model_response(
  data,
  fixed_effects_formula,
  fixed_effects_coefficients,
  random_effects_sd = NULL,
  random_effects_corr_matrix = list(),
  distribution = "gaussian",
  ...
)
}
\arguments{
\item{data}{A data frame with the original data used to generate the model matrix.}

\item{fixed_effects_formula}{A formula specifying the model.}

\item{fixed_effects_coefficients}{A vector of fixed effects coefficients.}

\item{random_effects_sd}{A named list of standard deviations for each random effect (default is NULL).}

\item{random_effects_corr_matrix}{A named list of correlation matrices for each random effect (default is an empty list).}

\item{distribution}{A character string specifying the distribution ("gaussian", "binomial", "poisson", "gamma").}

\item{...}{Additional parameters for the distribution functions.}
}
\value{
A vector representing the simulated response.
}
\description{
This function simulates effects based on the user's choice of distribution.
}
