% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{needs_sorting}
\alias{needs_sorting}
\title{Check if Names in a Vector Need Sorting to Match Names in a Data Frame}
\usage{
needs_sorting(vec, df)
}
\arguments{
\item{vec}{A named vector whose names are to be checked.}

\item{df}{A data frame whose column names are to be checked against the vector names.}
}
\value{
A logical value indicating whether the names in the vector need to be sorted to match the order in the data frame.
}
\description{
This function checks if the names in a given vector need to be sorted to match the order of names in a data frame.
}
