% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nest_data}
\alias{nest_data}
\title{Nest Data by Specified Columns}
\usage{
nest_data(data, nest_cols, data_col = "data")
}
\arguments{
\item{data}{A data frame or tibble to be nested.}

\item{nest_cols}{A character vector of column names to nest by.}

\item{data_col}{A character string specifying the name of the column to store the nested data. Defaults to "data".}
}
\value{
A tibble with nested data.
}
\description{
This function nests a data frame by specified columns.
}
