% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{random_filter_by_group}
\alias{random_filter_by_group}
\title{Filter and Sample Data by Groups}
\usage{
random_filter_by_group(data, ..., sample_size, drop_fails = FALSE)
}
\arguments{
\item{data}{A data frame to be filtered and sampled.}

\item{...}{Columns to group by.}

\item{sample_size}{An integer specifying the number of samples to take from each group.}

\item{drop_fails}{Boolean that, if TRUE, will drop any columns that don't have enough samples}
}
\value{
A data frame that has been filtered and sampled by the specified groups.
}
\description{
This function filters and samples a data frame by specified groups.
}
