% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-animals.R
\name{simulate_dams}
\alias{simulate_dams}
\title{Simulate Dams}
\usage{
simulate_dams(dams_per_group, ..., batches = 1, simulations = 1)
}
\arguments{
\item{dams_per_group}{A numeric vector indicating the number of dams per group.}

\item{...}{Additional arguments representing biological groups.}

\item{batches}{An integer indicating the number of batches.}

\item{simulations}{An integer indicating the number of simulations.}
}
\value{
A data frame containing the simulated dams.
}
\description{
This function simulates dams based on the provided parameters.
}
