% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-animals.R
\name{simulate_ideal_litters}
\alias{simulate_ideal_litters}
\title{Simulate Ideal Litters}
\usage{
simulate_ideal_litters(dams, ..., offspring_per_litter_group = 1)
}
\arguments{
\item{dams}{A data frame containing the dams.}

\item{...}{Additional arguments representing biological groups.}

\item{offspring_per_litter_group}{An integer indicating the number of offspring per litter group.}
}
\value{
A data frame containing the simulated offspring.
}
\description{
This function simulates ideal litters based on the provided parameters.
}
