% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{simulate_pups_response}
\alias{simulate_pups_response}
\title{Simulate Pups Response}
\usage{
simulate_pups_response(
  formula,
  fixed_effects_coefficients,
  dams_per_group,
  ...,
  simulations = 1,
  batches = 1,
  ideal = TRUE,
  offspring_per_litter_group = 1,
  mean_litter_size = 6,
  min_litter_size = 1,
  max_litter_size = 16,
  pup_characteristics = list(sex = c(female = 1, male = 1)),
  random_effects_sd = list(),
  random_effects_corr_matrix = list(),
  distribution = "gaussian",
  sigma = 1,
  weights = NULL,
  dispersion = NULL
)
}
\arguments{
\item{formula}{The formula for the model response.}

\item{fixed_effects_coefficients}{A named vector of fixed effects coefficients.}

\item{dams_per_group}{Number of dams per group.}

\item{...}{Additional arguments passed to `simulate_dams`.}

\item{simulations}{Number of simulations.}

\item{batches}{Number of batches.}

\item{ideal}{Logical indicating whether to simulate ideal litters.}

\item{offspring_per_litter_group}{Number of offspring per litter group.}

\item{mean_litter_size}{Mean litter size.}

\item{min_litter_size}{Minimum litter size.}

\item{max_litter_size}{Maximum litter size.}

\item{pup_characteristics}{List of pup characteristics.}

\item{random_effects_sd}{List of random effects standard deviations.}

\item{random_effects_corr_matrix}{List of random effects correlation matrices.}

\item{distribution}{Distribution type for the model response.}

\item{sigma}{Standard deviation of the residuals.}

\item{weights}{Weights for the model response.}

\item{dispersion}{Dispersion parameter for the model response.}
}
\value{
A data frame with the simulated pups and their responses.
}
\description{
This function simulates the response of pups based on the provided parameters.
}
