% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-animals.R
\name{simulate_real_litters}
\alias{simulate_real_litters}
\title{Simulate Real Litters}
\usage{
simulate_real_litters(
  dams,
  mean_litter_size = 6,
  ...,
  min_litter_size = 1,
  max_litter_size = 16
)
}
\arguments{
\item{dams}{A data frame containing the dams.}

\item{mean_litter_size}{A numeric value indicating the mean litter size.}

\item{...}{Additional arguments representing biological groups.}

\item{min_litter_size}{An integer indicating the minimum litter size.}

\item{max_litter_size}{An integer indicating the maximum litter size.}
}
\value{
A data frame containing the simulated offspring.
}
\description{
This function simulates real litters based on the provided parameters.
}
