% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-animals.R
\name{simulate_repeated_measurements}
\alias{simulate_repeated_measurements}
\title{Simulate Repeated Measurements}
\usage{
simulate_repeated_measurements(
  animals,
  replicates,
  grouping_var,
  id_col = "animal",
  column_name = "replicate"
)
}
\arguments{
\item{animals}{A data frame containing the animals.}

\item{replicates}{Either a numeric value (integer or double) or a character vector indicating the replicates.}

\item{grouping_var}{A character string indicating the grouping variable.}

\item{id_col}{A character string indicating the ID column name.}

\item{column_name}{A character string indicating the column name for replicates.}
}
\value{
A data frame containing the expanded data with repeated measurements.
}
\description{
This function simulates repeated measurements based on the provided parameters.
}
